+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- The comment expand/collapse indicators now match the Lobsters UI
+
 ## [1.64.0] - 2026-05-26
 
 ### Added